+++ app.js
@@ -1,2 +1,9 @@
+//require configuration
+requirejs.config({
+    baseUrl: "app",
+    paths: {
+        data: '../data'
+    }
+});
 // calling main function
-requirejs(["app/main"]);
+requirejs(["main"]);
