+++ app/viewmode/createSVG.js
@@ -1,0 +1,25 @@
+//create svg container w/ marker in a html container
+define(function(){
+    return function(container){
+
+        var width = 300,
+            height = 300;
+
+        var svg = d3.select(container).append("svg");
+
+        // marker creation
+        svg.append("defs")
+            .append("marker")
+            .attr("id","end")
+            .attr("viewBox","0 -5 10 10")
+            .attr("refX",25)
+            .attr("refY",-1)
+            .attr("markerWidth",6)
+            .attr("markerHeight",6)
+            .attr("orient","auto")
+            .append("path")
+            .attr("d","M0,-5L10,0L0,5");
+
+        return svg;
+    }
+})
